+++ angular/packages/angular-form-utils/src/public-api.ts
@@ -4,6 +4,6 @@ export type { FormDirtyState } from './lib/form-dirty-state';
 export { aggregateFormErrors, asyncFieldValidator } from './lib/form-errors';
 export { injectFormArrayDirtyState, arrayToggleItem, moveArrayItem, syncArrayValues } from './lib/form-array';
 export type { FormArrayDirtyState } from './lib/form-array';
-export { controlSignal, isControlInvalid, formDiff } from './lib/form-control-utils';
+export { controlSignal, isControlInvalid, formDiff, formStatusSignal, formSubmitHandler } from './lib/form-control-utils';
 export { IsInvalidPipe, FormErrorPipe } from './lib/form-pipes';
 export { ShowFormErrorDirective } from './lib/form-directives';